work on porting edge lengths to 64-bit

+++ src/redistribution/dgtog_calc_cnt.h
@@ -12,38 +12,38 @@ namespace CTF_int {
    * \brief computes the cardinality of the set of elements of a tensor of order idim+1 that are owned by processor index gidx_off in a distribution with dimensions sphase
    */
   template <int idim>
-  int64_t calc_cnt(int const * sym,
-                   int const * rep_phase,
-                   int const * sphase,
-                   int const * gidx_off,
-                   int const * edge_len,
-                   int const * loc_edge_len);
+  int64_t calc_cnt(int const *     sym,
+                   int const *     rep_phase,
+                   int const *     sphase,
+                   int const *     gidx_off,
+                   int64_t const * edge_len,
+                   int64_t const * loc_edge_len);
 
   /**
    * \brief computes the cardinality of the sets of elements of a tensor of order idim+1 for different values of the idim'th tensor dimension
    */
   template <int idim>
-  int64_t * calc_sy_pfx(int const * sym,
-                        int const * rep_phase,
-                        int const * sphase,
-                        int const * gidx_off,
-                        int const * edge_len,
-                        int const * loc_edge_len);
+  int64_t * calc_sy_pfx(int const *     sym,
+                        int const *     rep_phase,
+                        int const *     sphase,
+                        int const *     gidx_off,
+                        int64_t const * edge_len,
+                        int64_t const * loc_edge_len);
 
   template <int idim>
-  void calc_drv_cnts(int         ndim,
-                     int const * sym,
-                     int64_t *   counts,
-                     int const * rep_phase,
-                     int const * rep_phase_lda,
-                     int const * sphase,
-                     int const * phys_phase,
-                     int       * gidx_off,
-                     int const * edge_len,
-                     int const * loc_edge_len);
+  void calc_drv_cnts(int             ndim,
+                     int const *     sym,
+                     int64_t *       counts,
+                     int const *     rep_phase,
+                     int const *     rep_phase_lda,
+                     int const *     sphase,
+                     int const *     phys_phase,
+                     int       *     gidx_off,
+                     int64_t const * edge_len,
+                     int64_t const * loc_edge_len);
 
   void calc_drv_displs(int const *          sym,
-                       int const *          edge_len,
+                       int64_t const *      edge_len,
                        distribution const & old_dist,
                        distribution const & new_dist,
                        int64_t *            counts,
